create Track component
