+++ src/components/Tracks.js
@@ -1,9 +1,10 @@
 import React, { useEffect } from 'react';
 import axios from 'axios';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import Logout from './Logout';
 import Navigation from './Navigation';
 import { userTracks } from '../redux/actions/actions';
+import Track from './Track';
 
 const Tracks = () => {
   const dispatch = useDispatch();
@@ -20,43 +21,6 @@ const Tracks = () => {
     });
   };
 
-  const showTracks = () => {
-    const trackStored = useSelector((state) => state.tracks);
-    if (trackStored && trackStored.payload.length > 0) {
-      return trackStored.payload.map((track) => {
-        const {
-          day, electrodomestic, id, watts,
-        } = track;
-        const timeConnected = track.time_connected;
-        return (
-          <div key={id}>
-            <p>
-              Date:
-              {day}
-            </p>
-            <p>
-              Electrodomestic:
-              {electrodomestic}
-            </p>
-            <p>
-              Time Connected:
-              {timeConnected}
-            </p>
-            <p>
-              Watts:
-              {watts}
-            </p>
-          </div>
-        );
-      });
-    }
-    return (
-      <div>
-        <p> </p>
-      </div>
-    );
-  };
-
   useEffect(() => {
     fetchData();
   }, []);
@@ -64,7 +28,7 @@ const Tracks = () => {
   return (
     <div>
       <h1>Tracks</h1>
-      {showTracks()}
+      <Track />
       <Navigation />
       <Logout />
     </div>
